requirements.txt
numpy==1.26.4
pandas==2.2.3
datasets==3.1.0
tensorflow==2.10.0
uvicorn==0.32.0
fastapi==0.115.5
pillow==11.0.0
python-multipart==0.0.17